microservices/GeminiProcessing/requirements.txt
pika==1.2.1
python-docx
PyPDF2
ollama
requests
google-generativeai
python-dotenv